fix(fcp-raptorq): start repair ESI at K', remove asupersync dependency, harden chunk and config
SystematicEncoder: repair ESI cursor now starts at K' (smallest
K' >= K from the systematic index table) instead of K.  ESIs in
[K, K') are virtual padding and must not be emitted as repair
symbols.  All doc comments, assertions, and tests updated to
reflect K' semantics.

decoder.rs: Fix potential u32 overflow in repair_count calculation
by widening to u64 before multiplication.

config.rs: source_symbols() now uses u32::try_from().unwrap_or(u32::MAX)
instead of bare cast for payload_len division.

chunk.rs: Add assert!(chunk_size > 0) panic guard on from_payload().

Cargo.toml: Remove direct asupersync path dependency — raptorq
crate no longer needs it directly.

golden.rs: Minor formatting.

+++ crates/fcp-raptorq/src/codec/decoder.rs
@@ -2607,7 +2607,8 @@ mod tests {
 
         let mut transmitted: Vec<(u32, Vec<u8>)> =
             (0..k).map(|i| (i as u32, source_data[i].clone())).collect();
-        let repair_count = (k as u32 * repair_ratio_bps / 10_000).max(1);
+        let repair_count =
+            (u64::from(k as u32) * u64::from(repair_ratio_bps) / 10_000).max(1) as u32;
         for i in 0..repair_count {
             let esi = k_prime as u32 + i;
             transmitted.push((esi, encoder.repair_symbol(esi)));
